+++ OpenSim/Common/Component.cpp
@@ -166,8 +166,15 @@ void Component::prependComponentPathToConnecteePath(
     const Component& root = subcomponent.getRoot();
     for (auto& comp : subcomponent.updComponentList()) {
         for (auto& it : comp._socketsTable) {
-            if (!root.hasComponent(it.second->getConnecteePath()))
+            // Only apply prepend logic if the socket connection is within the
+            // added subcomponent. To do this, check if the connection is *not*
+            // available in the root component. For list sockets, we assume that
+            // checking the first connectee path is sufficient, since all
+            // connectees in the list must be from the same component.
+            if (it.second->getNumConnectees() > 0 &&
+                    !root.hasComponent(it.second->getConnecteePath(0))) {
                 it.second->prependComponentPathToConnecteePath(compPath);
+            }
         }
         for (auto& it : comp._inputsTable) {
             it.second->prependComponentPathToConnecteePath(compPath);